requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-pubsub==2.28.0
googleapis-common-protos==1.68.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
importlib_metadata==8.5.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
